+++ src/index.test.ts
@@ -219,6 +219,22 @@ test('Optional values', async () => {
   expect(output.errors).toStrictEqual({});
 });
 
+test('Branded values', async () => {
+  const schema = z.object({
+    name: z.string().brand('name')
+  });
+
+  const data = new FormData();
+  data.append('name', 'Name');
+
+  const output = await superValidate(data, schema);
+
+  expect(output.valid).equals(true);
+  expect(output.message).toBeUndefined();
+  expect(output.data.name).equals('Name');
+  expect(output.errors).toStrictEqual({});
+});
+
 describe('Default values', () => {
   test('With a partial entity', async () => {
     const now = new Date();