+++ src/lib/client/formEnhance.ts
@@ -13,7 +13,11 @@ import {
 import type { z, AnyZodObject } from 'zod';
 import { stringify } from 'devalue';
 import type { Entity } from '../schemaEntity.js';
-import type { FormOptions, SuperForm } from './index.js';
+import {
+  defaultOnError,
+  type FormOptions,
+  type SuperForm
+} from './index.js';
 import { clientValidation, validateField } from './clientValidation.js';
 import { Form } from './form.js';
 import { onDestroy } from 'svelte';
@@ -130,28 +134,38 @@ export function formEnhance<T extends AnyZodObject, M>(
   // Using this type in the function argument causes a type recursion error.
   const errors = errs as SuperForm<T, M>['errors'];
 
-  async function validateChange(
+  async function updateCustomValidity(
+    validityEl: HTMLElement,
+    event: 'blur' | 'input',
+    errors: string[] | undefined
+  ) {
+    if (!options.customValidity) return;
+    if (options.validationMethod == 'submit-only') return;
+
+    // Always reset validity, in case it has been validated on the server.
+    if ('setCustomValidity' in validityEl) {
+      (validityEl as HTMLInputElement).setCustomValidity('');
+    }
+
+    if (event == 'input' && options.validationMethod == 'onblur') return;
+
+    // If event is input but element shouldn't use custom validity,
+    // return immediately since validateField don't have to be called
+    // in this case, validation is happening elsewhere.
+    if (noCustomValidityDataAttribute in validityEl.dataset) return;
+
+    setCustomValidity(validityEl as HTMLInputElement, errors);
+  }
+
+  // Called upon an event from a HTML element that affects the form.
+  async function htmlInputChange(
     change: string[],
     event: 'blur' | 'input',
-    validityEl: HTMLElement | null
+    target: HTMLElement | null
   ) {
     if (options.validationMethod == 'submit-only') return;
 
-    if (options.customValidity && validityEl) {
-      // Always reset validity, in case it has been validated on the server.
-      if ('setCustomValidity' in validityEl) {
-        (validityEl as HTMLInputElement).setCustomValidity('');
-      }
-
-      if (event == 'input' && options.validationMethod == 'onblur') return;
-
-      // If event is input but element shouldn't use custom validity,
-      // return immediately since validateField don't have to be called
-      // in this case, validation is happening elsewhere.
-      if (noCustomValidityDataAttribute in validityEl.dataset)
-        if (event == 'input') return;
-        else validityEl = null;
-    }
+    //console.log('htmlInputChange', change, event, target);
 
     const result = await validateField(
       change,
@@ -161,29 +175,30 @@ export function formEnhance<T extends AnyZodObject, M>(
       tainted
     );
 
-    if (validityEl) {
-      setCustomValidity(validityEl as HTMLInputElement, result.errors);
-    }
+    // Update data if target exists (immediate is set, refactor please)
+    if (result.data && target) data.set(result.data);
 
-    // NOTE: Uncomment if Zod transformations should be immediately applied, not just when submitting.
-    // Not enabled because it's not great UX, and it's rare to have transforms, which will just result in
-    // redundant store updates.
-    //if (result.data) data.set(result.data);
+    if (options.customValidity) {
+      const name = CSS.escape(mergePath(change));
+      const el = formEl.querySelector<HTMLElement>(`[name="${name}"]`);
+      if (el) updateCustomValidity(el, event, result.errors);
+    }
   }
+
+  const immediateInputTypes = ['checkbox', 'radio', 'range'];
 
   /**
    * Some input fields have timing issues with the stores, need to wait in that case.
    */
-  function timingIssue(el: EventTarget | null) {
+  function isImmediateInput(el: EventTarget | null) {
     return (
       el &&
       (el instanceof HTMLSelectElement ||
         (el instanceof HTMLInputElement &&
-          (el.type == 'radio' || el.type == 'checkbox')))
+          immediateInputTypes.includes(el.type)))
     );
   }
 
-  // Add blur event, to check tainted
   async function checkBlur(e: Event) {
     if (
       options.validationMethod == 'oninput' ||
@@ -192,27 +207,24 @@ export function formEnhance<T extends AnyZodObject, M>(
       return;
     }
 
-    if (timingIssue(e.target)) {
-      await new Promise((r) => setTimeout(r, 0));
+    // Wait for changes to update
+    const immediateUpdate = isImmediateInput(e.target);
+    if (immediateUpdate) await new Promise((r) => setTimeout(r, 0));
+
+    const changes = get(lastChanges);
+    if (!changes.length) return;
+
+    const target = e.target instanceof HTMLElement ? e.target : null;
+
+    for (const change of changes) {
+      htmlInputChange(change, 'blur', immediateUpdate ? null : target);
     }
 
-    for (const change of get(lastChanges)) {
-      let validityEl: HTMLElement | null = null;
-
-      if (options.customValidity) {
-        const name = CSS.escape(mergePath(change));
-        validityEl = formEl.querySelector<HTMLElement>(`[name="${name}"]`);
-      }
-
-      validateChange(change, 'blur', validityEl);
-    }
     // Clear last changes after blur (not after input)
     lastChanges.set([]);
   }
-  formEl.addEventListener('focusout', checkBlur);
 
-  // Add input event, for custom validity
-  async function checkCustomValidity(e: Event) {
+  async function checkInput(e: Event) {
     if (
       options.validationMethod == 'onblur' ||
       options.validationMethod == 'submit-only'
@@ -220,34 +232,46 @@ export function formEnhance<T extends AnyZodObject, M>(
       return;
     }
 
-    if (timingIssue(e.target)) {
-      await new Promise((r) => setTimeout(r, 0));
-    }
+    // Wait for changes to update
+    const immediateUpdate = isImmediateInput(e.target);
+    if (immediateUpdate) await new Promise((r) => setTimeout(r, 0));
 
-    for (const change of get(lastChanges)) {
-      const name = CSS.escape(mergePath(change));
-      const validityEl = formEl.querySelector<HTMLElement>(
-        `[name="${name}"]`
-      );
-      if (!validityEl) continue;
+    const changes = get(lastChanges);
+    if (!changes.length) return;
 
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      const hadErrors = traversePath(get(errors), change as any);
-      if (hadErrors && hadErrors.key in hadErrors.parent) {
+    const target = e.target instanceof HTMLElement ? e.target : null;
+
+    for (const change of changes) {
+      const hadErrors =
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        immediateUpdate || traversePath(get(errors), change as any);
+      if (
+        immediateUpdate ||
+        (typeof hadErrors == 'object' && hadErrors.key in hadErrors.parent)
+      ) {
         // Problem - store hasn't updated here with new value yet.
-        setTimeout(() => validateChange(change, 'input', validityEl), 0);
+        setTimeout(
+          () =>
+            htmlInputChange(
+              change,
+              'input',
+              immediateUpdate ? target : null
+            ),
+          0
+        );
       }
     }
   }
 
-  if (options.customValidity) {
-    formEl.addEventListener('input', checkCustomValidity);
-  }
+  formEl.addEventListener('focusout', checkBlur);
+  formEl.addEventListener('input', checkInput);
 
   onDestroy(() => {
     formEl.removeEventListener('focusout', checkBlur);
-    formEl.removeEventListener('input', checkCustomValidity);
+    formEl.removeEventListener('input', checkInput);
   });
+
+  ///// SvelteKit enhance function //////////////////////////////////
 
   const htmlForm = Form(formEl, { submitting, delayed, timeout }, options);
 
@@ -284,13 +308,15 @@ export function formEnhance<T extends AnyZodObject, M>(
       if (options.flashMessage) cancelFlash(options);
     } else {
       // Client validation
-      const validation = await clientValidation(
-        formEl.noValidate ||
+      const noValidate =
+        !options.SPA &&
+        (formEl.noValidate ||
           ((submit.submitter instanceof HTMLButtonElement ||
             submit.submitter instanceof HTMLInputElement) &&
-            submit.submitter.formNoValidate)
-          ? undefined
-          : options.validators,
+            submit.submitter.formNoValidate));
+
+      const validation = await clientValidation(
+        noValidate ? undefined : options.validators,
         get(data),
         get(formId),
         get(constraints),
@@ -401,7 +427,14 @@ export function formEnhance<T extends AnyZodObject, M>(
     }
 
     async function validationResponse(event: ValidationResponse) {
-      const result = event.result;
+      // Check if an error was thrown in hooks, in which case it has no type.
+      const result: ActionResult = event.result.type
+        ? event.result
+        : {
+            type: 'error',
+            status: 500,
+            error: event.result
+          };
 
       currentRequest = null;
       let cancelled = false;
@@ -483,8 +516,14 @@ export function formEnhance<T extends AnyZodObject, M>(
             if (options.onError !== 'apply') {
               const data = { result, message };
 
-              for (const event of formEvents.onError) {
-                if (event !== 'apply') await event(data);
+              for (const onErrorEvent of formEvents.onError) {
+                if (
+                  onErrorEvent !== 'apply' &&
+                  (onErrorEvent != defaultOnError ||
+                    !options.flashMessage?.onError)
+                ) {
+                  await onErrorEvent(data);
+                }
               }
             }
           }