+++ src/lib/schemaEntity.ts
@@ -20,7 +20,8 @@ import type {
   ZodNumber,
   ZodDate,
   ZodArray,
-  ZodPipeline
+  ZodPipeline,
+  ZodBranded
 } from 'zod';
 
 import type { SuperValidateOptions } from './superValidate.js';
@@ -97,23 +98,34 @@ export function unwrapZodType(zodType: ZodTypeAny): ZodTypeInfo {
   //let i = 0;
   while (_wrapped) {
     //console.log(' '.repeat(++i * 2) + zodType.constructor.name);
-    if (zodType._def.typeName == 'ZodNullable') {
-      isNullable = true;
-      zodType = (zodType as ZodNullable<ZodTypeAny>).unwrap();
-    } else if (zodType._def.typeName == 'ZodDefault') {
-      hasDefault = true;
-      defaultValue = zodType._def.defaultValue();
-      zodType = zodType._def.innerType;
-    } else if (zodType._def.typeName == 'ZodOptional') {
-      isOptional = true;
-      zodType = (zodType as ZodOptional<ZodTypeAny>).unwrap();
-    } else if (zodType._def.typeName == 'ZodEffects') {
-      if (!effects) effects = zodType as ZodEffects<ZodTypeAny>;
-      zodType = zodType._def.schema;
-    } else if (zodType._def.typeName == 'ZodPipeline') {
-      zodType = (zodType as ZodPipeline<ZodTypeAny, ZodTypeAny>)._def.out;
-    } else {
-      _wrapped = false;
+    switch (zodType._def.typeName) {
+      case 'ZodNullable':
+        isNullable = true;
+        zodType = (zodType as ZodNullable<ZodTypeAny>).unwrap();
+        break;
+      case 'ZodDefault':
+        hasDefault = true;
+        defaultValue = zodType._def.defaultValue();
+        zodType = zodType._def.innerType;
+        break;
+      case 'ZodOptional':
+        isOptional = true;
+        zodType = (zodType as ZodOptional<ZodTypeAny>).unwrap();
+        break;
+      case 'ZodEffects':
+        if (!effects) effects = zodType as ZodEffects<ZodTypeAny>;
+        zodType = zodType._def.schema;
+        break;
+      case 'ZodPipeline':
+        zodType = (zodType as ZodPipeline<ZodTypeAny, ZodTypeAny>)._def.out;
+        break;
+      case 'ZodBranded':
+        zodType = (
+          zodType as ZodBranded<ZodTypeAny, string | number | symbol>
+        ).unwrap();
+        break;
+      default:
+        _wrapped = false;
     }
   }